requirements.txt
opencv-python==4.8.1.78
mediapipe==0.10.8
numpy==1.26.2
Flask==3.0.0
pyserial==3.5
pygame==2.5.2
